Update Poseta na MENDO.cpp

+++ Programs/Mendo/Poseta na MENDO.cpp
@@ -9,7 +9,7 @@ int main() {
     cin.tie(0);
     int k, n;
     cin >> k >> n;
-    ll has[2][101] = {};
+    long long has[2][101] = {};
     for (int j = 0; j < 2; ++j){
         for (int i = 0; i < n; ++i) {
             int x;
@@ -17,9 +17,9 @@ int main() {
             ++has[j][x];
         }
     }
-    ll ans = 0;
+    long long ans = 0;
     for (int a = 1; a <= 100; ++a) {
-        for (int b=1; b <= 100; ++b) {
+        for (int b = 1; b <= 100; ++b) {
             if (a+b >= k && has[0][a] && has[1][b]) {
                 ans += has[0][a] * has[1][b];
             }
